refactor: extract CalibrationCalculator service and CameraMixin (#112)
